chore: split dr tests, some test qol
